SDK-61 refactor some tests
